pyproject.toml
[project]
name = "tapie-react-board-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "tortoise-orm[asyncpg]>=0.25.0",
    "uvicorn>=0.34.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
